コントローラー設定

+++ app/controllers/recipes_controller.rb
@@ -1,7 +1,7 @@
 class RecipesController < ApplicationController
   before_action :authenticate_user!, except: [:index, :show]
   before_action :set_recipe, only: [:show, :edit]
-  before_action :move_to_index, only: [:edit, :update, :destroy]
+  before_action :move_to_index, only: [:edit, :update]
 
   def index
     @recipes = Recipe.order("created_at DESC")
@@ -37,10 +37,16 @@ class RecipesController < ApplicationController
 
   def destroy
     recipe = Recipe.find(params[:id])
-    recipe.destroy
-    redirect_to root_path
+    if current_user.id == recipe.user_id
+      recipe.destroy
+      redirect_to root_path
+    end
   end
 
+  def category
+    @q = Recipe.ransack(params[:q])
+    @recipes = @q.result
+  end
 
   private
